jwt funs() added, userlogin, logout added

+++ resources/item.py
@@ -1,5 +1,10 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    jwt_optional,
+    get_jwt_claims,
+    fresh_jwt_required
+)
 from models.item_model import ItemModel
 
 
@@ -12,19 +17,21 @@ class Item(Resource):
         required=True,
         help="This field cannot be left blank!"
     )
+
     parser.add_argument('store_id',
         type=int,
         required=True,
         help="Every item needs a store_id"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}
@@ -36,22 +43,25 @@ class Item(Resource):
             return {"message": "An error occurred inserting the item."}
         return item.json()
 
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {'message': 'Item deleted'}
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
-        if item is None:
-            item = ItemModel(name, **data)
+        if item:
+            item.price = data['price']
+            item.store_id = data['store_id']
         else:
-            item['price'] = data['price']
-            item['store_id'] = data['store_id']
+            item = ItemModel(name, **data)
         item.save_to_db()
         return item.json()
 
@@ -59,5 +69,11 @@ class Item(Resource):
 class ItemList(Resource):
     TABLE_NAME = 'items'
 
+    @jwt_optional
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        user_id = get_jwt_claims()
+        items = [x.json() for x in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}
+        return {'items': [item['name'] for item in items],
+                'message': "More data available if you log in."}
